Modified to use replaceState instead of pushState
It's great to have the history, but most of the time I want to come back from a page and show the most recent sort/filter that I had on the grid.  I don't want to sort a grid and then click the back button, only to go back to the previous sort.