Fixed bug on init where bindToolbarEvents() was called before reloadGrid finished
This is actually kind of a serious bug because this is called when the
DOM is in a half-way state as it is reloading.  I have a change stashed
that i will be checking in after this that uncovered it in a very
apparent way.  if you call setURLAndReload before the bindToolbarEvents
in this init function, the additional parameters and filters will not be
set in the UI because they have not been rendered on the DOM yet.